refactor: pr review 테스트 코드의 이름을 명시적으로 변경하라
