+++ src/test/java/nextstep/subway/line/acceptance/LineAcceptanceSupport.java
@@ -12,7 +12,7 @@ import nextstep.subway.line.dto.LineResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-public class LineAcceptance {
+public class LineAcceptanceSupport {
     public static ExtractableResponse<Response> 지하철_노선_등록되어_있음(LineRequest params) {
         return 지하철_노선_생성_요청(params);
     }
