+++ src/test/java/nextstep/subway/line/acceptance/LineAcceptanceTest.java
@@ -12,18 +12,18 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_등록되어_있음;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_목록_응답됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_목록_조회_요청;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_목록_포함됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_삭제됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_생성_실패됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_생성_요청;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_생성됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_수정_요청;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_수정됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_응답됨;
-import static nextstep.subway.line.acceptance.LineAcceptance.지하철_노선_제거_요청;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_목록_응답됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_목록_조회_요청;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_목록_포함됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_삭제됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_생성_실패됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_생성_요청;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_생성됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_수정_요청;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_수정됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_응답됨;
+import static nextstep.subway.line.acceptance.LineAcceptanceSupport.지하철_노선_제거_요청;
 
 @DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
