+++ src/test/java/nextstep/subway/line/acceptance/SectionAcceptanceSupport.java
@@ -13,7 +13,7 @@ import nextstep.subway.station.dto.StationResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-public class LineSectionAcceptance {
+public class SectionAcceptanceSupport {
     public static ExtractableResponse<Response> 지하철_노선에_지하철역_등록_요청(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
         SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
